+++ e2e/e2etests/test_eth_deposit_call.go
@@ -1,0 +1,132 @@
+package e2etests
+
+import (
+	"fmt"
+	"math/big"
+	"strings"
+
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	ethtypes "github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/zeta-chain/zetacore/e2e/runner"
+	"github.com/zeta-chain/zetacore/e2e/utils"
+	testcontract "github.com/zeta-chain/zetacore/testutil/contracts"
+	"github.com/zeta-chain/zetacore/x/crosschain/types"
+)
+
+// TestEtherDepositAndCall tests deposit of ethers calling a example contract
+func TestEtherDepositAndCall(r *runner.E2ERunner, args []string) {
+	if len(args) != 1 {
+		panic("TestEtherDepositAndCall requires exactly one argument for the amount.")
+	}
+
+	value, ok := big.NewInt(0).SetString(args[0], 10)
+	if !ok {
+		panic("Invalid amount specified for TestEtherDepositAndCall.")
+	}
+
+	r.Logger.Info("Deploying example contract")
+	exampleAddr, _, exampleContract, err := testcontract.DeployExample(r.ZEVMAuth, r.ZEVMClient)
+	if err != nil {
+		panic(err)
+	}
+	r.Logger.Info("Example contract deployed")
+
+	// preparing tx
+	evmClient := r.EVMClient
+	gasLimit := uint64(23000)
+	gasPrice, err := evmClient.SuggestGasPrice(r.Ctx)
+	if err != nil {
+		panic(err)
+	}
+	nonce, err := evmClient.PendingNonceAt(r.Ctx, r.DeployerAddress)
+	if err != nil {
+		panic(err)
+	}
+
+	data := append(exampleAddr.Bytes(), []byte("hello sailors")...)
+	tx := ethtypes.NewTransaction(nonce, r.TSSAddress, value, gasLimit, gasPrice, data)
+	chainID, err := evmClient.NetworkID(r.Ctx)
+	if err != nil {
+		panic(err)
+	}
+	deployerPrivkey, err := crypto.HexToECDSA(r.DeployerPrivateKey)
+	if err != nil {
+		panic(err)
+	}
+	signedTx, err := ethtypes.SignTx(tx, ethtypes.NewEIP155Signer(chainID), deployerPrivkey)
+	if err != nil {
+		panic(err)
+	}
+
+	r.Logger.Info("Sending a cross-chain call to example contract")
+	err = evmClient.SendTransaction(r.Ctx, signedTx)
+	if err != nil {
+		panic(err)
+	}
+	receipt := utils.MustWaitForTxReceipt(r.Ctx, r.EVMClient, signedTx, r.Logger, r.ReceiptTimeout)
+	if receipt.Status == 0 {
+		panic("tx failed")
+	}
+	cctx := utils.WaitCctxMinedByInTxHash(r.Ctx, signedTx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
+	if cctx.CctxStatus.Status != types.CctxStatus_OutboundMined {
+		panic(fmt.Sprintf("expected cctx status to be mined; got %s", cctx.CctxStatus.Status))
+	}
+
+	// Checking example contract has been called, bar value should be set to amount
+	bar, err := exampleContract.Bar(&bind.CallOpts{})
+	if err != nil {
+		panic(err)
+	}
+	if bar.Cmp(value) != 0 {
+		panic(fmt.Sprintf("cross-chain call failed bar value %s should be equal to amount %s", bar.String(), value.String()))
+	}
+	r.Logger.Info("Cross-chain call succeeded")
+
+	r.Logger.Info("Deploying reverter contract")
+	reverterAddr, _, _, err := testcontract.DeployReverter(r.ZEVMAuth, r.ZEVMClient)
+	if err != nil {
+		panic(err)
+	}
+	r.Logger.Info("Example reverter deployed")
+
+	// preparing tx for reverter
+	gasPrice, err = evmClient.SuggestGasPrice(r.Ctx)
+	if err != nil {
+		panic(err)
+	}
+	nonce, err = evmClient.PendingNonceAt(r.Ctx, r.DeployerAddress)
+	if err != nil {
+		panic(err)
+	}
+
+	data = append(reverterAddr.Bytes(), []byte("hello sailors")...)
+	tx = ethtypes.NewTransaction(nonce, r.TSSAddress, value, gasLimit, gasPrice, data)
+	signedTx, err = ethtypes.SignTx(tx, ethtypes.NewEIP155Signer(chainID), deployerPrivkey)
+	if err != nil {
+		panic(err)
+	}
+
+	r.Logger.Info("Sending a cross-chain call to reverter contract")
+	err = evmClient.SendTransaction(r.Ctx, signedTx)
+	if err != nil {
+		panic(err)
+	}
+
+	receipt = utils.MustWaitForTxReceipt(r.Ctx, r.EVMClient, signedTx, r.Logger, r.ReceiptTimeout)
+	if receipt.Status == 0 {
+		panic("tx failed")
+	}
+
+	cctx = utils.WaitCctxMinedByInTxHash(r.Ctx, signedTx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
+	if cctx.CctxStatus.Status != types.CctxStatus_Reverted {
+		panic(fmt.Sprintf("expected cctx status to be reverted; got %s", cctx.CctxStatus.Status))
+	}
+	r.Logger.Info("Cross-chain call to reverter reverted")
+
+	// check the status message contains revert error hash in case of revert
+	// 0xbfb4ebcf is the hash of "Foo()"
+	if !strings.Contains(cctx.CctxStatus.StatusMessage, "0xbfb4ebcf") {
+		panic(fmt.Sprintf("expected cctx status message to contain revert reason; got %s", cctx.CctxStatus.StatusMessage))
+	}
+}
