+++ e2e/e2etests/test_zeta_deposit_new_address.go
@@ -5,19 +5,21 @@ import (
 
 	"github.com/zeta-chain/zetacore/e2e/runner"
 	"github.com/zeta-chain/zetacore/e2e/utils"
+	"github.com/zeta-chain/zetacore/testutil/sample"
 )
 
-func TestERC20Deposit(r *runner.E2ERunner, args []string) {
+func TestZetaDepositNewAddress(r *runner.E2ERunner, args []string) {
 	if len(args) != 1 {
-		panic("TestERC20Deposit requires exactly one argument for the amount.")
+		panic("TestZetaDepositNewAddress requires exactly one argument for the amount.")
 	}
 
 	amount, ok := big.NewInt(0).SetString(args[0], 10)
 	if !ok {
-		panic("Invalid amount specified for TestERC20Deposit.")
+		panic("Invalid amount specified for TestZetaDepositNewAddress.")
 	}
 
-	hash := r.DepositERC20WithAmountAndMessage(r.DeployerAddress, amount, []byte{})
+	newAddress := sample.EthAddress()
+	hash := r.DepositZetaWithAmount(newAddress, amount)
 
 	// wait for the cctx to be mined
 	cctx := utils.WaitCctxMinedByInTxHash(r.Ctx, hash.Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
